Clean up. Seperate backend and frontend. Check db connection before running server
