+++ backend/config.js
@@ -1,0 +1,1 @@
+export const PORT = 8080;
